docker compose

+++ docker-demo/src/main/java/com/demo/dockerdemo/model/Film.java
@@ -24,5 +24,5 @@ public class Film {
   private String description;
 
   @Column(name  = "language_id")
-  private int languageId;
+  private long languageId;
 }
